Remove interface with only one implementation
