+++ app/Error/ErrorHandlerProvider.php
@@ -46,7 +46,7 @@ class ErrorHandlerProvider implements ServiceProviderInterface
 
                 $logger->pushHandler(new SentryHandler($app[HubInterface::class], Logger::ERROR));
 
-                return new PsrLoggerErrorHandler(
+                return new ErrorHandler(
                     $logger,
                     $app['jwt'] ?? null,
                     $app['auth.api_key'] ?? null,
